feat(form): support configuring the AJV verification delay (no delay by default)
