+++ packages/hooks/src/index.ts
@@ -14,7 +14,7 @@ export { default as useRefProp } from './useRefProp'
 export { default as usePrevious } from './usePrevious'
 export { default as useEventCallback } from './useEventCallback'
 export { default as useClickOne } from './useClickOne'
-export { default as useValidate } from './useValidate'
+export { default as useValidate, validate } from './useValidate'
 export { default as useField } from './useField'
 export { default as useQuery } from './useQuery'
 export { default as useArray } from './useArray'
